Handle project changes from the server
This handles the custom $/command notification from the server and stores
the settings in the project_data() of the Window. If the project data is not
backed by a .sublime-project, then a dialog will appear that tells the user
to save the project, otherwise the changes are lost the next time the folder
is opened.